rename: ChallengeParticipation 참여 관련 예외 클래스를 participation 패키지로 이동
