+++ module-api/src/main/java/com/trybe/moduleapi/challenge/exception/participation/NotFoundChallengeParticipationException.java
@@ -1,4 +1,4 @@
-package com.trybe.moduleapi.challenge.exception;
+package com.trybe.moduleapi.challenge.exception.participation;
 
 import com.trybe.moduleapi.common.api.exception.BusinessException;
 import org.springframework.http.HttpStatus;
